test: semantically splits admin e2e
